requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
fastapi
uvicorn[standard]
pydantic
